Factored out source reading and source injection for SGs, proper root namespaces
