+++ Sources/SynKit/Libraries/Directory.Build.props
@@ -5,5 +5,8 @@
     <DocumentationFile>$(MSBuildProjectDirectory)/bin/doc/Yoakke.SynKit.$(MSBuildProjectName).xml</DocumentationFile>
     <AssemblyName>Yoakke.SynKit.$(MSBuildProjectName)</AssemblyName>
     <PackageId>Yoakke.SynKit.$(MSBuildProjectName)</PackageId>
+
+    <!-- Root namespace -->
+    <RootNamespace>Yoakke.SynKit.$(MSBuildProjectName)</RootNamespace>
   </PropertyGroup>
 </Project>
