+++ Sources/SynKit/Tests/Directory.Build.props
@@ -1,3 +1,8 @@
 <Project>
   <Import Project="..\..\Props\Test.Build.props" />
+
+  <PropertyGroup>
+    <!-- Root namespace -->
+    <RootNamespace>Yoakke.SynKit.$(MSBuildProjectName)</RootNamespace>
+  </PropertyGroup>
 </Project>
